Clase usada para comprimir string
I am trying to compress a String. 
For example, if the user input is "aaabbcccd" - the output should count the letters and if count is above 1, print the letter, 
then the number: a3b2c3d1.

if the String to be compressed is small than the result String it may return the string to be compressed as result.